third_party/WebKit/LayoutTests/virtual/user-activation-v2/user-activation-v2/full-screen-gamepad-expected.txt

EXPECTED (internals.runtimeFlags.userActivationV2Enabled == 'true') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLHtmlElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLSpanElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLHtmlElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == 'undefined') OK
END OF TEST
